server/requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.2.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==24.3.0
blinker==1.9.0
Brotli==1.0.9
cachetools==5.3.3
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.3
Flask==3.1.0
Flask-Cors==3.0.10
flatbuffers==24.3.25
frozenlist==1.5.0
gast==0.4.0
google-auth==2.29.0
google-auth-oauthlib==0.5.2
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.12.1
idna==3.7
itsdangerous==2.2.0
jax==0.4.30
jaxlib==0.4.30
Jinja2==3.1.5
keras==2.12.0
Keras-Preprocessing==1.1.2
libclang==18.1.1
Markdown==3.4.1
MarkupSafe==2.1.3
ml_dtypes==0.5.1
multidict==6.1.0
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==24.2
pillow==11.0.0
pip==24.2
propcache==0.2.0
protobuf==3.20.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.10.1
pyOpenSSL==23.2.0
PySocks==1.7.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.7.2
scipy==1.15.1
setuptools==72.1.0
six==1.16.0
tensorboard==2.12.1
tensorboard_data_server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-hub==0.8.0
termcolor==2.1.0
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
wheel==0.35.1
wrapt==1.14.1
yarl==1.18.0


gunicorn==23.0.0 